Create Visualize.ipynb
This class contains a number of functions for dimensionality reduction and visualization of
higher-dimensional datasets. It also contains a function to train a K-Means model on the 
input data. The number of clusters that the K-Means model uses can be informed by the graphs 
generated by the plot_2D and plot_3D functions - these can show the number of clusters in the 
data. The dimensionality reduction functions, Reduce_Dim_PCA and Reduce_Dim_SVD, can be used 
to reduce higher-dimensional datasets to 2D or 3D datasets that can then be graphed, for the 
purpose of determining the number of clusters in the data, as outlined above.